+++ android/src/main/kotlin/com/mjohnsullivan/flutterwear/wear/WearPlugin.kt
@@ -1,90 +1,162 @@
 package com.mjohnsullivan.flutterwear.wear
 
+import android.os.Bundle
+import androidx.core.view.ViewCompat.getRootWindowInsets
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.LifecycleObserver
+import androidx.lifecycle.OnLifecycleEvent
+import com.google.android.wearable.compat.WearableActivityController
+import io.flutter.embedding.engine.plugins.FlutterPlugin
+import io.flutter.embedding.engine.plugins.activity.ActivityAware
+import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
+import io.flutter.embedding.engine.plugins.lifecycle.HiddenLifecycleReference
+import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
-import io.flutter.plugin.common.MethodCall
-import io.flutter.plugin.common.PluginRegistry.Registrar
-import io.flutter.view.FlutterView
 
-import android.app.Activity
-import android.os.Bundle
-
-import android.support.v4.view.ViewCompat.requestApplyInsets
-import android.support.v4.view.ViewCompat.setOnApplyWindowInsetsListener
-import android.support.v4.view.WindowInsetsCompat
-
-import android.support.wear.ambient.AmbientMode
 
 private const val CHANNEL_NAME = "wear"
 
-fun getChannel(view: FlutterView): MethodChannel {
-    return MethodChannel(view, CHANNEL_NAME)
-}
-
-class WearPlugin private constructor(
-        activity: Activity,
-        private val view: FlutterView,
-        val channel: MethodChannel) :
-        MethodCallHandler {
-
-    private var mAmbientController: AmbientMode.AmbientController? = null
-
-    init {
-        // Set the Flutter ambient callbacks
-        mAmbientController = AmbientMode.attachAmbientSupport(activity)
-    }
+class WearPlugin : FlutterPlugin, ActivityAware, MethodCallHandler, LifecycleObserver, WearableActivityController.AmbientCallback() {
+    private var mMethodChannel: MethodChannel? = null
+    private var mActivityBinding: ActivityPluginBinding? = null
+    private var mAmbientController: WearableActivityController? = null
 
     companion object {
-        @JvmStatic
-        fun registerWith(registrar: Registrar) {
-            val channel = MethodChannel(registrar.messenger(), CHANNEL_NAME)
-            channel.setMethodCallHandler(WearPlugin(registrar.activity(),
-                    registrar.view(), channel))
-        }
+        const val TAG = "WearPlugin"
+        const val BURN_IN_PROTECTION = WearableActivityController.EXTRA_BURN_IN_PROTECTION
+        const val LOWBIT_AMBIENT = WearableActivityController.EXTRA_LOWBIT_AMBIENT
     }
 
+    override fun onAttachedToEngine(binding: FlutterPlugin.FlutterPluginBinding) {
+        mMethodChannel = MethodChannel(binding.binaryMessenger, CHANNEL_NAME)
+        mMethodChannel!!.setMethodCallHandler(this)
+    }
 
+    override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {
+        mMethodChannel?.setMethodCallHandler(this)
+        mMethodChannel = null
+    }
+
+    override fun onAttachedToActivity(binding: ActivityPluginBinding) {
+        attachAmbientController(binding)
+    }
+
+    override fun onDetachedFromActivityForConfigChanges() {
+        detachAmbientController()
+    }
+
+    override fun onReattachedToActivityForConfigChanges(binding: ActivityPluginBinding) {
+        attachAmbientController(binding)
+    }
+
+    override fun onDetachedFromActivity() {
+        detachAmbientController()
+    }
+
+    private fun attachAmbientController(binding: ActivityPluginBinding) {
+        mActivityBinding = binding
+        mAmbientController = WearableActivityController(TAG, binding.activity, this)
+        mAmbientController?.setAmbientEnabled();
+        val reference = (binding.lifecycle as HiddenLifecycleReference)
+        reference.lifecycle.addObserver(this)
+    }
+
+    private fun detachAmbientController() {
+        mActivityBinding?.let {
+            val reference = (it.lifecycle as HiddenLifecycleReference)
+            reference.lifecycle.removeObserver(this)
+        }
+        mActivityBinding = null
+
+    }
 
     override fun onMethodCall(call: MethodCall, result: Result) {
-        when {
-            call.method == "getPlatformVersion" ->
-                result.success("Android ${android.os.Build.VERSION.RELEASE}")
-            call.method == "getShape" -> handleShapeMethodCall(result)
+        when (call.method) {
+            "getShape" -> {
+                val view = mActivityBinding?.activity?.window?.decorView
+                when {
+                    view == null -> {
+                        result.error("no-view", "No android view available.", null)
+                    }
+                    getRootWindowInsets(view)?.isRound == true -> {
+                        result.success("round")
+                    }
+                    else -> {
+                        result.success("square")
+                    }
+                }
+            }
+            "isAmbient" -> {
+                result.success(mAmbientController?.isAmbient ?: false)
+            }
+            "setAutoResumeEnabled" -> {
+                val enabled = call.argument<Boolean>("enabled")
+                if (mAmbientController == null || enabled == null) {
+                    result.error("not-ready", "Ambient mode controller not ready", null)
+                } else {
+                    mAmbientController!!.setAutoResumeEnabled(enabled)
+                    result.success(null)
+                }
+            }
+            "setAmbientOffloadEnabled" -> {
+                val enabled = call.argument<Boolean>("enabled")
+                if (mAmbientController == null || enabled == null) {
+                    result.error("not-ready", "Ambient mode controller not ready", null)
+                } else {
+                    mAmbientController!!.setAmbientOffloadEnabled(enabled)
+                    result.success(null)
+                }
+            }
             else -> result.notImplemented()
         }
     }
 
-    private fun handleShapeMethodCall(result: Result) {
-        setOnApplyWindowInsetsListener(view, { _, insets: WindowInsetsCompat? ->
-            if (insets?.isRound == true) {
-                result.success(0)
-            } else {
-                result.success(1)
-            }
-            WindowInsetsCompat(insets)
-        })
-        requestApplyInsets(view)
-    }
-}
-
-/*
- * Pass ambient callback back to Flutter
- */
-class FlutterAmbientCallback(private val channel: MethodChannel) : AmbientMode.AmbientCallback() {
-
     override fun onEnterAmbient(ambientDetails: Bundle) {
-        channel.invokeMethod("enter", null)
-        super.onEnterAmbient(ambientDetails)
+        val burnInProtection = ambientDetails.getBoolean(BURN_IN_PROTECTION, false)
+        val lowBitAmbient = ambientDetails.getBoolean(LOWBIT_AMBIENT, false)
+        mMethodChannel?.invokeMethod("onEnterAmbient", mapOf(
+                "burnInProtection" to burnInProtection,
+                "lowBitAmbient" to lowBitAmbient
+        ))
     }
 
     override fun onExitAmbient() {
-        channel.invokeMethod("exit", null)
-        super.onExitAmbient()
+        mMethodChannel?.invokeMethod("onExitAmbient", null)
     }
 
     override fun onUpdateAmbient() {
-        channel.invokeMethod("update", null)
-        super.onUpdateAmbient()
+        mMethodChannel?.invokeMethod("onUpdateAmbient", null)
     }
+
+    override fun onInvalidateAmbientOffload() {
+        mMethodChannel?.invokeMethod("onInvalidateAmbientOffload", null)
+    }
+
+    @OnLifecycleEvent(Lifecycle.Event.ON_CREATE)
+    fun onCreate() {
+        mAmbientController?.onCreate()
+    }
+
+    @OnLifecycleEvent(Lifecycle.Event.ON_RESUME)
+    fun onResume() {
+        mAmbientController?.onResume()
+    }
+
+    @OnLifecycleEvent(Lifecycle.Event.ON_PAUSE)
+    fun onPause() {
+        mAmbientController?.onPause()
+    }
+
+    @OnLifecycleEvent(Lifecycle.Event.ON_STOP)
+    fun onStop() {
+        mAmbientController?.onStop()
+    }
+
+    @OnLifecycleEvent(Lifecycle.Event.ON_DESTROY)
+    fun onDestroy() {
+        mAmbientController?.onDestroy()
+    }
+
 }