+++ example/android/app/src/main/kotlin/com/mjohnsullivan/flutterwear/wear_example/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.mjohnsullivan.flutterwear.wear_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}