+++ ios/Classes/WearPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface WearPlugin : NSObject<FlutterPlugin>
-@end