+++ ios/Classes/WearPlugin.m
@@ -1,8 +1,0 @@
-#import "WearPlugin.h"
-#import <wear/wear-Swift.h>
-
-@implementation WearPlugin
-+ (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar>*)registrar {
-  [SwiftWearPlugin registerWithRegistrar:registrar];
-}
-@end