+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## [0.0.1] - TODO: Add release date.
+# 0.1.0
 
-* TODO: Describe initial release.
+- Updated to AndroidX and Android embedding v2.
+- Renamed `Shape` is now `WearShape`.
+- Renamed `Mode` is now `WearMode`.
+- Deprecated `InheritedShape`.
+  _Add `WatchShape` instead and use `WatchShape.of(context)` to get the shape value._
+
+# 0.0.1
+
+- Initial release